delete article function, mysql table update
1. delete article function
2. change fastjson version
3.enable fastjson
4. fix comment page bug (level number error)
5. delete foreign keys for all tables in mysql

+++ cbse-blog-server/src/main/java/com/loloao/service/ArticleService.java
@@ -31,7 +31,7 @@ public interface ArticleService extends IService<Article> {
 
     Long updateArticle(Article article);
 
-    void deleteArticleById(Integer id);
+    void deleteArticleById(Long id);
 
     List<Article> listArticlesByTag(Integer id);
